updated marshal and added exchange-database advertiseAs options.
advertiseAs is necessary if you want to deal internally with entities
that contain sensible information and use a sibling class for the public area.
Classic use-case is a User and PublicUser, where both extend BaseUse.
User has the password field which PublicUser doesn't have. So, internally
you use User, and when using ExchangeDatabase you use `advertiseAs: FrontendUser`,
so the password is not accidentally published.