added instance pooling, new Query methods: updateOne, patchOne, patchMany, deleteOne, deleteMany.
with this refactor the Database API changed dramatically, and it is required to do all
modification and fetching via the Database.query(Class) api.
Use weak-napi to detect when a entity instance is not used anymore to remove it automatically
from the instance pool.
Added more tests and updated typedoc.