Refactored marshal API to better support (nested) container types
Before this commit it was not possible to either have nested types (string[][]), or attach
validators to the container type (array/map) directly. With this commit this is
now possible + it aligns the decorator API a bit to not force execute decorator properties
all the time.

Also renamed `f` to `t` everywhere (alias to `type`), and deprecate `f`.

See also #53.

Examples:

```typescript
@t.optional() -> @t.optional

@t.primary().uuid() -> @t.primary.uuid

@t.decorated() -> @t.decorated

@t.asName('foo') -> @t.name('foo')

@t.any().asArray() -> @t.array(t.any)

@t.enum(MyEnum).asArray() -> @t.array(t.enum(MyEnum))

@t.any() -> @t.any

@t.forward(() => User) -> @t.type(() => User)

@t.array(String).optional() -> @t.array(t.string.optional)

//supports now container validation and content types validation
@t.array(String).validator(MinStringLength(3)) ->
@t.array(t.string.optional.validator(MinStringLength(3))).validator(MinArrayLength(3))

@t.array(t.array(t.string))
tags: string[][]
```