removed type checking validating for object mismatch
Since TypeScript's reflection system is too weak, we had to remove that check entirely.
TyepScript reports for `x | undefined` a type of `Object`, which is obviously wrong.
But we used `Object` to check for invalid type definition, which is not valid always.
It hurts more than it helps in UX so removed it entirely.

fixes #69