refactor(http): change the way parameter resolver work and make them much faster.
The old version relied on a method `supports` which is called for each parameter
for each registered resolver during runtime. Now, this is not needed anymore
and we detect during build-time which resolver should be used for which
route parameter.

Also fixed an issue with having multiple resolvers on the same route.