feature(sql): add raw SQL support
With
  database.query(T).where(sql`id > 23`).find()
and
  database.raw(sql`SELECT * FROM ${T} id > 23`).find()

it now very easy to execute raw SQL queries.

For generic Database<any> a new function `isDatabaseOf` is added
to allow type guarding for SQLDatabaseAdapter database objects.

  if (isDatabaseOf(database, SQLDatabaseAdapter)) {
    const items = await database.raw(sql`SELECT * ...`).find();
  }

with `identifier` can column names be quoted automatically to the
right syntax depending on the adapter used.

  const name = 'peter';
  database.raw(sql`SELECT * FROM ${User} where ${identifier('name')} = ${name}`).find()

Values added as ${} to the TS template of `sql` are automatically added to the
prepared statement as placeholders.