feature(type): allow to use references in plain objects. feature(orm): performance improvements.
With
  plainToClass(Blog, {author: 2})
`author` is now deserialized as a Reference if it is @t.reference().

With:
  plainToClass(Blog, {author: database.getReference(User, 2)})
the reference is kept and no error is thrown.

Moved more reference (hydrated&hydrator) information from @deepkit/orm to @deepkit/type to have it better separated.

Separated ORM's InstanceState and class state JIT functions to gain a small performance improvement. Resetting the ChangeSet instead of setting a new one improves fetching times further.