feature(orm): add Query.disableChangeDetection()
This allows to disable change-detection for returned objects. It's not always necessary to build a change-detection snapshot for each object when the user is not even modifying the object and (perhaps) only returns the result directly to the client. In this case, they can disable the change-detection behavior which results in a nice performance boost.