feature(http): express/connect-like middleware support
With

  Application.create({
    middlewares: [
        middleware.for(compress()).forController(MyHttpController),
        middleware.for(MyMiddleware).forRoutes({path: '/api/*'}),
    ]
  }).run();

or

    @http.middleware(MyMiddleware)
    class MyControllerA {
        @http.GET('/b/:name').middleware(AnotherMiddleware)
        route1(name: string) {
            return name;
        }
    }

a user can now register arbitrary middlewares passed to `.for(T)` with the signature of

   (request: HttpRequest, response: HttpResponse, next: (error?: any) => void): void;

or
   class MyMiddleware {
       execute(request: HttpRequest, response: HttpResponse, next: (error?: any) => void) {
           response.setHeader('middleware_a', '1');
           next();
       }
   }

and thus supports almost all express middlewares. `httpMiddleware.for(T)` has many chain methods
that allow to filter further to define where the middleware is allowed be executed.
All middlewares are executed in the `onRoute` workflow place, right after the route is resolved.

Also, a new in-memory request handling API is added that allows much easier to
test the http kernel. e.g.

    const response = await httpKernel.request(HttpRequest.GET('/user/name1').query({v: 23}));
    expect(response.statusCode).toEqual(200);
    expect(response.bodyString).toEqual('"name1"');
    expect(response.getHeader('content-type')).toEqual('application/json; charset=utf-8');

Also, when a content-type header is set, it's not overwritten in onResponse http kernel anymore.