feature(orm): transaction support
Added two APIs to allow using transactions for Mongo, SQLite, MySQL, and Postgres.

    const session = database.createSession();
    session.useTransaction(); //marks this session as transactional
    try {
        session.add(...);
        await session.query(...);

        await session.commit();
    } catch (error) {
        await session.rollback();
        throw error;
    }

and a less verbose way that commits automatically and rollbacks when the callback throws:

    await session.transaction(async (session) => {
        session.add(...);
        await session.query(...);
    });
The returned object of `session.useTransaction()` is adapter dependent
and enables for mysql/postgres additional transaction option like isolation level.

With this change, the DatabaseSession has a new method `flush`,
which is doing the same as `commit`, but without committing the transaction (if one is active).