feature(orm): deep (dot path) patch support for all database adapters.
```
await database.query(Order)
    .filter({id: 1})
    .patchOne({'shippingAddress.postCode': '12345'});
```

Reference updates are not supported this way, only embedded documents (that are stored for example via json/jsonb in MySQL/Postgres).

With this change a new deepkit/type function is available: patch (and getPatchSerializeFunction) which allows to serialize/deserialize a patch-like structure and rename properties at the same time.

e.g. `{'shippingAddress.street': 'abc'}` => `{'shipping_address.street': 'abc'}`

This functionality is necessary to convert TS types to database types in the ORM when `.patchMany({'shippingAddress.street': 'abc'})` is executed.