feature(sql): allow to exclude entities/tables from migration
With this change unknown entities will no longer be deleted in the database per default.
This means if an entity gets removed from the Database class, it's table won't be deleted automatically. The user has to do that manually.

With `@entity.excludeMigration()` is now possible
to stop creating schema migration for a particular entity.

```typescript
@entity.excludeMigration()
class UnmanagedEntity {
    id: number & PrimaryKey;
}
```

fix also issues with creating and reading migration files.