fix(injector): make sure type cache is used when finding matching provider.
Previously a check was wrongly implemented so that isExtendable got no a Type object and has to infer its type first. This happens without cache per default as the function is designed to operate on Type objects and if non is given it assumes it runs in the Processor (where caching for inferring types is disabled).

This commit also adds new debugging output when computation takes a very long time (>100ms)