fix(type-compiler): merge compilerOptions from ConfigResolver
Since not all build tools provide a proper compilerOptions object (with `paths`) from the real config file, we load it manually and apply its options to the current ReflectionTransformer.compilerOptions.

This way we have always the full picture of the compilerOptions including paths that is vital for path resolving in Resolver.