fix(type): handle `Function` correctly in type comparisons
Previously `Function` was not unique represented in the runtime type system and hence it was not possible to compare other types with it.
This is now fixed so that `Function` is in runtime types `{kind: Kind.Function, function: Function}` which can be detected via the global Function symbol `type.function === Function`.

Adjusts `extend` check accordingly and makes sure more function comparisons are supported.