feat(rpc): add http transport
This change refactors the transport layer of RPC in a way that allows to serialize RpcMessageDefinition in a custom way.
This is used then in a new `RpcHttpClientAdapter` adapter. This is a first experimental version of it.

This contains breaking changes in the way RPC adapters are written. RpcMessageWriter does not exist anymore and became an implementation detail. A new TransportClientConnection and TransportConnection abstraction has been added that is used in both: the client and server.